==> 0-element_replacement.ipynb <==
import pandas as pd
import numpy as np
import threading
import pandas as pd
import math

A_cations = ['Sm','Sc','Gd','Y','La','Pr','Nd','Tm','Rb','Dy','Pm','Yb',
             'Mg','Hf','Cs','Ce','Ca','Pb','Cd','Bi','Ba','Tl','Ag','Tb',
             'Sb','Eu','K','Sn','In','Na','Sr']
B_cations = ['Cr','Co','Fe','Nb','In','Er','Sn','Zr','Be','Ge','Hf','Ti','Ga','Si','Ta','Al','Bi','B','Sb','Ce']

X_anions = ['O']
A_ion_radii = {
    'Eu':1.437,'La':1.36,'Pr':1.304,'Nd':1.27,'Tm':1.168,'Rb':1.72,'Dy':1.192,'Pm':1.339,'Yb':1.343,'Mg':1.21,'Hf':1.07,
    'Cs':1.88,'Gd':1.212,'Ce':1.14,'Ca':1.34,'Pb':1.49,'Ba':1.61,'Sr':1.44,'Cd':1.31,'Y':1.223,'Sm':1.24,'Tl':1.70,'Bi':1.380,
    'In':1.596,'Ag':1.521,'Tb':1.201,'Sc':1.122,'Sb':1.234,'K':1.64,'Sn':1.408,'Na':1.39
}
B_ion_radii = {
    'Cr':0.615,'Nb':0.64,'In':0.80,'Er':0.89,'Sn':0.69,'Zr':0.72,'Be':0.45,'Ge':0.53,'Ti':0.605,'Si':0.40,'Sn':0.69,'Fe':0.645,
    'Co':0.61,'Sb':0.60,'Ta':0.64,'Al':0.535,'Bi':1.03,'B':0.27,'Ce':0.87,'Hf':0.71,'Ga':0.62
}
X_ion_radii = {'O':1.40}
nA_jiatai = {
    'Eu':2,'La':3,'Pr':3,'Nd':3,'Tm':3,'Rb':1,'Dy':3,'Pm':3,'Yb':2,'Mg':2,'Hf':4,'Cs':1,'Gd':3,'Ce':4,'Ca':2,'Pb':2,
    'Ba':2,'Cd':2,'Ag':1,'Tb':3,'Sc':3,'Sb':3,'K':1,'Sn':2,'Na':1,'Sr':2,'Cd':2,'Y':3,'Sm':3,'Tl':1,'Bi':3,'In':1
}

# no new_tolerance

# All possible chemical formulas after ion substitution
import concurrent.futures
from pymatgen.core import Composition
def generate_candidate_all(i,site):
    candidates = []
    for X in X_anions:
        if site == 'A':
            for A1 in A_cations:
                for A2 in A_cations:
                    for B in B_cations:
                        B1, B2 = B,B
                        b = 1
                        a = i
                        if (A1 != A2) and (A1 != B) and (A2 != B):
                            candidate = ''.join([X, '3', A1, str(a), A2, str(1 - a), B1, str(b), B2, str(1 - b)])
                            tmp_dict = {'formula': Composition(candidate).reduced_formula,
                                        'A1': A1, 'A1_ratio': a,'nA1':nA_jiatai[A1],
                                        'A2': A2, 'A2_ratio': 1 - a,'nA2':nA_jiatai[A2],
                                        'B1': B1, 'B1_ratio': b, 'B2': B2, 'B2_ratio': 1 - b,
                                        'X': X, 'X_ratio': 3,
                                        'Ra': A_ion_radii[A1] * a + A_ion_radii[A2] * (1 - a),
                                        'Rb': B_ion_radii[B1] * b + B_ion_radii[B2] * (1 - b),
                                        'Rx': X_ion_radii[X],
                                        'nA': round(nA_jiatai[A1] * a + nA_jiatai[A2] * (1 - a))}
                            candidates.append(tmp_dict)
        else:
            for A in A_cations:
                for B1 in B_cations:
                    for B2 in B_cations:
                        A1, A2 = A,A
                        a = 1
                        b = i 
                        if (B1 != B2) and (B1 != A) and (B2 != A):
                            candidate = ''.join([X, '3', A1, str(a), A2, str(1 - a), B1, str(b), B2, str(1 - b)])
                            tmp_dict = {'formula': Composition(candidate).reduced_formula,
                                        'A1': A1, 'A1_ratio': a,'nA1':nA_jiatai[A1],
                                        'A2': A2, 'A2_ratio': 1 - a,'nA2':nA_jiatai[A2],
                                        'B1': B1, 'B1_ratio': b, 'B2': B2, 'B2_ratio': 1 - b,
                                        'X': X, 'X_ratio': 3,
                                        'Ra': A_ion_radii[A1] * a + A_ion_radii[A2] * (1 - a),
                                        'Rb': B_ion_radii[B1] * b + B_ion_radii[B2] * (1 - b),
                                        'Rx': X_ion_radii[X],
                                        'nA': nA_jiatai[A1]}
                            candidates.append(tmp_dict)
    return candidates

def generate_candidates_concurrently_all(dope_range,site):
    all_candidates = []
    with concurrent.futures.ThreadPoolExecutor(max_workers=128) as executor:
        # Create all possible combinations of a and b values
        futures = [executor.submit(generate_candidate_all, a,site) 
                   for a in dope_range]
        for future in concurrent.futures.as_completed(futures):
            all_candidates.extend(future.result())
        result = pd.DataFrame(all_candidates)
    return result.drop_duplicates(subset=['formula'], keep='first').reset_index(drop=True)

# new_tolerance

# All possible chemical formulas after ion substitution (add tolerance factor screening)
import concurrent.futures
from pymatgen.core import Composition
def generate_candidate(i,site):
    candidates = []
    for X in X_anions:
        if site == 'A':
            for A1 in A_cations:
                for A2 in A_cations:
                    for B in B_cations:
                        B1, B2 = B,B
                        b = 1
                        a = i
                        if (A1 != A2) and (A1 != B) and (A2 != B):
                            candidate = ''.join([X, '3', A1, str(a), A2, str(1 - a), B1, str(b), B2, str(1 - b)])
                            tmp_dict = {'formula': Composition(candidate).reduced_formula,
                                        'A1': A1, 'A1_ratio': a,'nA1':nA_jiatai[A1],
                                        'A2': A2, 'A2_ratio': 1 - a,'nA2':nA_jiatai[A2],
                                        'B1': B1, 'B1_ratio': b, 'B2': B2, 'B2_ratio': 1 - b,
                                        'X': X, 'X_ratio': 3,
                                        'Ra': A_ion_radii[A1] * a + A_ion_radii[A2] * (1 - a),
                                        'Rb': B_ion_radii[B1] * b + B_ion_radii[B2] * (1 - b),
                                        'Rx': X_ion_radii[X],
                                        'nA': round(nA_jiatai[A1] * a + nA_jiatai[A2] * (1 - a))}
                            new_tor = tmp_dict['Rx']/tmp_dict['Rb'] - tmp_dict['nA']*(tmp_dict['nA'] - (tmp_dict['Ra']/tmp_dict['Rb'])/np.log1p(tmp_dict['Ra']/tmp_dict['Rb'] - 1))
                            if (tmp_dict['Ra'] > tmp_dict['Rb']) and (new_tor <= 4.18):
                                candidates.append(tmp_dict)
                            else:
                                continue
        else:
            for A in A_cations:
                for B1 in B_cations:
                    for B2 in B_cations:
                        A1, A2 = A,A
                        a = 1
                        b = i 
                        if (B1 != B2) and (B1 != A) and (B2 != A):
                            candidate = ''.join([X, '3', A1, str(a), A2, str(1 - a), B1, str(b), B2, str(1 - b)])
                            tmp_dict = {'formula': Composition(candidate).reduced_formula,
                                        'A1': A1, 'A1_ratio': a,'nA1':nA_jiatai[A1],
                                        'A2': A2, 'A2_ratio': 1 - a,'nA2':nA_jiatai[A2],
                                        'B1': B1, 'B1_ratio': b, 'B2': B2, 'B2_ratio': 1 - b,
                                        'X': X, 'X_ratio': 3,
                                        'Ra': A_ion_radii[A1] * a + A_ion_radii[A2] * (1 - a),
                                        'Rb': B_ion_radii[B1] * b + B_ion_radii[B2] * (1 - b),
                                        'Rx': X_ion_radii[X],
                                        'nA': nA_jiatai[A1]}
                            new_tor = tmp_dict['Rx']/tmp_dict['Rb'] - tmp_dict['nA']*(tmp_dict['nA'] - (tmp_dict['Ra']/tmp_dict['Rb'])/np.log1p(tmp_dict['Ra']/tmp_dict['Rb'] - 1))
                            if (tmp_dict['Ra'] > tmp_dict['Rb']) & (new_tor <= 4.18):
                                candidates.append(tmp_dict)
                            else:
                                continue
    return candidates

def generate_candidates_concurrently(dope_range,site):
    all_candidates = []
    with concurrent.futures.ThreadPoolExecutor(max_workers=128) as executor:
        # Create all possible combinations of a and b values
        futures = [executor.submit(generate_candidate, a,site) 
                   for a in dope_range]
        for future in concurrent.futures.as_completed(futures):
            all_candidates.extend(future.result())
        result = pd.DataFrame(all_candidates)
    return result.drop_duplicates(subset=['formula'], keep='first').reset_index(drop=True)

dope_range = np.arange(0, 1.1, 0.1)
site = 'A'
result_a_all = generate_candidates_concurrently_all(dope_range,site)
result_a_all

result_b_all = generate_candidates_concurrently_all(dope_range,'B')
result_b_all

dope_range = np.arange(0, 1.1, 0.1)
site = 'A'
result_a_2 = generate_candidates_concurrently(dope_range,site)
result_a_2

# result_a_2.to_csv('/home/zhouhairui/paper_code/a_replace_data.csv')

result_b_2 = generate_candidates_concurrently(dope_range,'B')
result_b_2

# result_b_2.to_csv('/home/zhouhairui/paper_code/b_replace_data.csv')

# result_b.to_csv('b_dop_data_2.csv')

df_2 = pd.concat([result_a_2, result_b_2],axis = 0).reset_index(drop=True)
df_2

# df_2.to_csv('/home/zhouhairui/paper_code/replacement_data.csv')
