==> 2-optimize_bandgap.ipynb <==
import pandas as pd
import numpy as np
from pymatgen.core import Composition
import re
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib.pyplot as plt
from matplotlib import rcParams
from IPython.core.pylabtools import figsize
import seaborn as sns
from sklearn.model_selection import cross_val_score
import warnings
warnings.filterwarnings("ignore")

band_gap = pd.read_csv('/data/zhr_bandgap.csv')
band_gap= band_gap.drop(columns = 'Unnamed: 0')
band_gap

from sklearn.cluster import KMeans
kmeans = KMeans(n_clusters=2, random_state=0)
def exclude_diff_hsedata(tmp:pd.Series):
    if np.sum(np.array(tmp) == 0) >= 4:
        return 0
    else:
        kmeans.fit(np.array(tmp).reshape(-1,1))
        labels = kmeans.labels_
        res = pd.DataFrame({'value': tmp, 'label': labels})
        class0 = res[res['label']==0]
        class1 = res[res['label']==1]
        if len(class0) > len(class1):
            return class0['value'].mean()
        else:
            return class1['value'].mean()

from time import time
start = time()
band_gap['mean'] = band_gap.iloc[:,2:7].apply(lambda x: exclude_diff_hsedata(x), axis=1)
end = time()
print(end - start)

# band_gap.to_csv('zhr_bandgap_correct.csv')
