==> 3-materias_classification.ipynb <==
import pandas as pd
import numpy as np
from pymatgen.core import Composition
import re
import seaborn as sns
import matplotlib.pyplot as plt
from matplotlib import rcParams
from IPython.core.pylabtools import figsize
import seaborn as sns
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import KFold
from sklearn.ensemble import RandomForestRegressor
from sklearn.ensemble import ExtraTreesRegressor
from sklearn.model_selection import train_test_split
from sklearn.feature_selection import SelectFromModel
from sklearn.model_selection import cross_val_score
from tqdm import tqdm
import warnings
warnings.filterwarnings('ignore')
from sklearn.metrics import accuracy_score
from sklearn.metrics import mean_squared_error
from sklearn.metrics import mean_absolute_error,mean_squared_error,r2_score
import warnings
warnings.filterwarnings("ignore")

small_dataset = pd.read_csv('/ata/single_perovskite_oxide.csv')
small_dataset = small_dataset.drop(columns = 'Unnamed: 0').reset_index(drop=True)
small_dataset = small_dataset.drop(columns = 'composition').reset_index(drop=True)
small_dataset['composition'] = small_dataset['formula'].map(Composition)
small_dataset

non_single_oxide = pd.read_csv('/data/non_single_oxide.csv').drop(columns = 'Unnamed: 0')

single_oxide = small_dataset.drop(columns = ['target','A','B','composition'])
non_single_oxide = non_single_oxide.drop(columns = 'target')

single_oxide['is_single_oxide'] = 1
non_single_oxide['is_single_oxide'] = 0
data = pd.concat([single_oxide,non_single_oxide],axis=0).reset_index(drop=True)
data

from matminer.featurizers.base import MultipleFeaturizer
from matminer.featurizers import composition as cf
feature_calculators = MultipleFeaturizer([cf.ElementProperty.from_preset("magpie"),
                                          cf.ValenceOrbital(props=['avg']), cf.IonProperty(fast=True)])
feature_labels = feature_calculators.feature_labels()
data['composition'] = data['formula'].map(Composition)
data_features = feature_calculators.featurize_dataframe(data,col_id='composition')

def remove_collinear_features(x, threshold):
    '''
    Objective:
        Remove collinear features in a dataframe with a correlation coefficient
        greater than the threshold. Removing collinear features can help a model
        to generalize and improves the interpretability of the model.
        
    Inputs: 
        threshold: any features with correlations greater than this value are removed
    
    Output: 
        dataframe that contains only the non-highly-collinear features
    '''  
    # Calculate the correlation matrix
    corr_matrix = x.corr()
    iters = range(len(corr_matrix.columns) - 1)
    drop_cols = []

    # Iterate through the correlation matrix and compare correlations
    for i in iters:
        for j in range(i):
            item = corr_matrix.iloc[j:(j+1), (i+1):(i+2)]
            col = item.columns
            row = item.index
            val = abs(item.values)
            
            # If correlation exceeds the threshold
            if val >= threshold:
                # Print the correlated features and the correlation value
                # print(col.values[0], "|", row.values[0], "|", round(val[0][0], 2))
                drop_cols.append(col.values[0])

    # Drop one of each pair of correlated columns
    drops = set(drop_cols)
    x = x.drop(columns = drops)
    return x

X = data_features[feature_labels]
y = data['is_single_oxide']
X = remove_collinear_features(X, 0.8)
feature_labels = X.columns

X_train,X_test,y_train,y_test = train_test_split(X,y,shuffle = True,random_state=42)
from sklearn.model_selection import StratifiedKFold
from tpot import TPOTClassifier
cv = StratifiedKFold(n_splits=5) 

model_clf = TPOTClassifier(generations=10,population_size=50,cv=cv, scoring='roc_auc', verbosity=2, 
                       random_state=42, n_jobs=-1) 
model_clf.fit(X_train, y_train)

best_model = model_clf.fitted_pipeline_

best_model.fit(X_train, y_train)

from sklearn.metrics import roc_auc_score
from sklearn.metrics import roc_auc_score,accuracy_score,f1_score,matthews_corrcoef,classification_report,roc_curve, auc,precision_score,recall_score
y_test_pred = best_model.predict(X_test)

acc = accuracy_score(y_test, y_test_pred)
print(f"accuracy: {acc}")
roc_auc = roc_auc_score(y_test,y_test_pred)
print(f"roc_auc: {roc_auc}")
f1 = f1_score(y_test, y_test_pred ,average='weighted')
print(f"f1: {f1}")

from joblib import dump
model_file_path = 'space_clf_tpot.joblib'
dump(best_model, model_file_path)
print("Model saved successfully at:", model_file_path)

import numpy as np
import pandas as pd
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.model_selection import train_test_split


X_train,X_test,y_train,y_test = train_test_split(X, y, random_state=42)

# Average CV score on the training set was: 0.9467857142857143
model_clf = GradientBoostingClassifier(learning_rate=0.1, max_depth=2, max_features=0.25, min_samples_leaf=9, min_samples_split=10, n_estimators=100, subsample=0.7500000000000001)
# Fix random state in exported estimator
if hasattr(model_clf, 'random_state'):
    setattr(model_clf, 'random_state', 42)

model_clf.fit(X_train, y_train)

# from sklearn.metrics import roc_auc_score
# y_test_pred = model_clf.predict(X_test)
# roc_auc = roc_auc_score(y_test,y_test_pred)
# roc_auc

a_dop_data = pd.read_csv('data/a_replace_data.csv').drop(columns = 'Unnamed: 0')
# b_dop_data = pd.read_csv('b_replace_data.csv').drop(columns = 'Unnamed: 0')

b_dop_data = pd.read_csv('data/b_replace_data.csv').drop(columns = 'Unnamed: 0')

b_dop_data

a_dop_data['composition'] = a_dop_data['formula'].map(Composition)
# b_dop_data['composition'] = b_dop_data['formula'].map(Composition)

b_dop_data['composition'] = b_dop_data['formula'].map(Composition)

a_data_features = feature_calculators.featurize_dataframe(a_dop_data, col_id='composition',ignore_errors=True)

a_data_features

a_dop_data = pd.merge(a_dop_data,a_data_features,on = 'formula')
a_dop_data = a_dop_data.drop_duplicates(subset=['formula'], keep='first').reset_index(drop=True)
a_dop_data

b_data_features = feature_calculators.featurize_dataframe(b_dop_data, col_id='composition',ignore_errors=True)

b_dop_data = pd.merge(b_dop_data,b_data_features,on = 'formula')
b_dop_data = b_dop_data.drop_duplicates(subset=['formula'], keep='first').reset_index(drop=True)
b_dop_data

a_X = a_data_features[feature_labels]
a_dop_data['is_metal_oxide'] = model_clf.predict(a_X)
a_dop_data['is_metal_oxide'].value_counts()

b_X = b_data_features[feature_labels]
b_dop_data['is_metal_oxide'] = model_clf.predict(b_X)
b_dop_data['is_metal_oxide'].value_counts()

a_data_221 = a_dop_data[a_dop_data['is_metal_oxide']==1]
b_data_221 = b_dop_data[b_dop_data['is_metal_oxide']==1]
print(a_data_221.shape,b_data_221.shape)

feature_calculators = MultipleFeaturizer([cf.ElementProperty.from_preset("magpie"),
                                          cf.ValenceOrbital(props=['avg']), cf.IonProperty(fast=True)])
magpie_feature_labels = feature_calculators.feature_labels()

dop_data_221 = pd.concat([a_data_221,b_data_221],axis=0).reset_index(drop = True)
dop_data_221

a_data_221

# dop_data_221.to_csv('dop_data_221_new.csv')

# a_data_221.to_csv('a_data_221_new.csv')
# b_data_221.to_csv('b_data_221_new.csv')
